fix: move TestCliContextArguments into cli_context_test.go
